muse-hardware/requirements.txt
Flask==3.1.1
Flask-SocketIO==5.5.1
Flask-CORS==5.0.0
numpy==1.26.4
opencv-python==4.10.0.82
onnx==1.18.0
pandas==2.3.1
psutil==7.0.0
python-socketio==5.13.0
gpiozero==2.0.5
lgpio==0.2.2.0
spacemit-ort==1.2.2
pillow==11.3.0
torch==2.5.0
torchvision==0.19.1
ultralytics==8.3.169
pyserial==3.5
pycryptodome==3.21.0
schedule==1.2.2
pymodbus==3.6.3